+++ modules/Fun.Files/Fun.Files.psm1
@@ -2,6 +2,8 @@
 . "$PSScriptRoot\public\Show-FileContents.ps1"
 . "$PSScriptRoot\public\Get-FileContents.ps1"
 . "$PSScriptRoot\public\Copy-FileContents.ps1"
+. "$PSScriptRoot\public\Compress-FilteredFiles.ps1"
 
-Export-ModuleMember `
-    -Function Show-FileContents, Get-FileContents, Invoke-FileTransform, Copy-FileContents
+Export-ModuleMember -Function `
+    Show-FileContents, Get-FileContents, Invoke-FileTransform, Copy-FileContents, `
+    Compress-FilteredFiles